Accept Types.OTHER knowing it's what we get for UUID
This is a pretty lame hack to accept values reflected with type OTHER
from the database schema. Types.OTHER is what it gets when the column
is defined as a UUID type. Using a string to store in kafka for a UUID
is totally fine. Only readers need to know about the fact it is a UUID
to handle the values read from kafka properly.